# Conflicts:
#	Assets/Prefabs/NewPlayerPrefabNetworking.prefab
#	Assets/Scenes/BelkaLobby.unity
#	Assets/Scenes/GabFluids.unity
#	Assets/Scripts/HealthController.cs

+++ Assets/Scripts/HealthController.cs
@@ -25,14 +25,14 @@ public class HealthController : NetworkBehaviour
     {
         applyDamage(collider);
     }
-
+    
     //when collision, apply damages to player's health component
     [Server]
     public void applyDamage(Collider collider)
     {
         if (collider.gameObject.tag.Contains("Drop"))
         {
-            int tmp_dmg = UnityEngine.Random.Range(100, 150);
+            float tmp_dmg = UnityEngine.Random.Range(100, 150);
             health.Health -= tmp_dmg; //TODO way of computing damage=f(power)?
             informations.log("<b><color=\"blue\">" + gameObject.name + "</color></b>: -" + tmp_dmg + "PV");
             if (health.Health < 1)
